+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Enh #875: Ignore "Packets out of order..." warnings in `AbstractPdoCommand::internalExecute()` method (@Tigrov)
 - Enh #877: Separate column type constants (@Tigrov)
 - Enh #878: Realize `ColumnBuilder` class (@Tigrov)
+- Enh #881: Refactor `ColumnSchemaInterface` and `AbstractColumnSchema` (@Tigrov) 
 
 ## 1.3.0 March 21, 2024